UI layer

+++ app/src/main/java/com/vlyashuk/bookcase/network/model/RetailPrice.kt
@@ -1,6 +1,6 @@
 package com.vlyashuk.bookcase.network.model
 
 data class RetailPrice(
-    val amountInMicros: Int,
+    val amountInMicros: Double,
     val currencyCode: String
 )
